requirements.txt
asgiref==3.7.2
Django==4.0.10
djangorestframework==3.13.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
